Restructure code to facilitate using build script
Completion for shells should be generated at compile time. In order for
this to happen, we need to use a build script, which would include our
code before we even compile it.

Therefore, in order to facilitate this premature inclusion, I have split
code into two crates, one for the CLI interface and the other for its
core, which then became a library.